refactor: address copilot comments
