+++ src/docs/index.ts
@@ -1,6 +1,6 @@
 export {
   OPENAPI_DOCS_PATH,
   OPENAPI_GLOBAL_PREFIX,
-  generateOpenApiFiles,
   registerOpenApi,
 } from './swagger.document';
+export { generateOpenApiFiles } from './swagger.generate';
