+++ src/docs/swagger.document.ts
@@ -1,10 +1,5 @@
 import { INestApplication } from '@nestjs/common';
-import { NestFactory } from '@nestjs/core';
 import { SwaggerModule } from '@nestjs/swagger';
-import { existsSync, mkdirSync, writeFileSync } from 'fs';
-import { join } from 'path';
-import * as yaml from 'js-yaml';
-import { AppModule } from '../app.module';
 import { createSwaggerConfig, getSwaggerMetadata } from './swagger.config';
 
 export const OPENAPI_GLOBAL_PREFIX = 'api/v1';
@@ -15,19 +10,9 @@ const documentOptions = {
   deepScanRoutes: true as const,
 };
 
-type OpenApiDocument = ReturnType<typeof createOpenApiDocument>;
+export type OpenApiDocument = ReturnType<typeof createOpenApiDocument>;
 
-function ensureOpenApiGenerationEnv(): void {
-  if (!process.env.NODE_ENV) {
-    process.env.NODE_ENV = 'test';
-  }
-  process.env.DATABASE_URL ??= 'postgresql://mock:mock@localhost:5432/mock_db';
-  process.env.KEYCLOAK_BASE_URL ??= 'http://localhost:8080';
-  process.env.KEYCLOAK_REALM ??= 'mock-realm';
-  process.env.KEYCLOAK_CLIENT_ID ??= 'mock-client-id';
-}
-
-function createOpenApiDocument(app: INestApplication) {
+export function createOpenApiDocument(app: INestApplication) {
   const config = createSwaggerConfig();
   return SwaggerModule.createDocument(app, config, documentOptions);
 }
@@ -36,6 +21,7 @@ function setupOpenApiDocs(app: INestApplication, document: OpenApiDocument): voi
   const { apiVersion, apiRelease } = getSwaggerMetadata();
 
   SwaggerModule.setup(OPENAPI_DOCS_PATH, app, document, {
+    useGlobalPrefix: false,
     swaggerOptions: {
       persistAuthorization: true,
       tagsSorter: 'alpha',
@@ -74,42 +60,7 @@ function setupOpenApiDocs(app: INestApplication, document: OpenApiDocument): voi
   });
 }
 
-function writeOpenApiArtifacts(document: OpenApiDocument, outputDir: string): void {
-  if (!existsSync(outputDir)) {
-    mkdirSync(outputDir, { recursive: true });
-  }
-
-  writeFileSync(
-    join(outputDir, 'openapi.json'),
-    JSON.stringify(document, null, 2),
-    'utf8',
-  );
-  writeFileSync(
-    join(outputDir, 'openapi.yaml'),
-    yaml.dump(document, { indent: 2 }),
-    'utf8',
-  );
-}
-
 export function registerOpenApi(app: INestApplication): void {
   const document = createOpenApiDocument(app);
   setupOpenApiDocs(app, document);
-}
-
-export async function generateOpenApiFiles(
-  outputDir: string,
-): Promise<OpenApiDocument> {
-  ensureOpenApiGenerationEnv();
-
-  const app = await NestFactory.create(AppModule, {
-    logger: false,
-    abortOnError: false,
-  });
-  app.setGlobalPrefix(OPENAPI_GLOBAL_PREFIX);
-
-  const document = createOpenApiDocument(app);
-  writeOpenApiArtifacts(document, outputDir);
-  await app.close();
-
-  return document;
 }
